Bot/requirements.txt
flask
python-dotenv
aiogram
httpx
requests
uvicorn
quart
hypercorn